backend/requirements.txt
django==3.2.0
djangorestframework
pillow
cors-headers